applied all requested changes in comments
