+++ README.md
@@ -350,7 +350,10 @@ test('playwright-graphql test negative', async ({ gql }) => {
 
 ### GraphQL Coverage Reporting
 
-Track and visualize which GraphQL operations and arguments have been tested:
+Track and visualize which GraphQL operations and arguments have been tested,
+complex input params with nested types and enums will be tracked and converted to readable format:
+
+![DEMO](docs/coverage-html.png)
 
 Playwright config file for the GraphQL coverage reporter:
 1. **graphqlFilePath** (required): Path to the autogenerated GraphQL file with getSdk .
@@ -367,7 +370,7 @@ import { defineConfig } from '@playwright/test';
 export default defineConfig({
     reporter: [
         ['list'],
-        ['playwright-graphql/reporter', {
+        ['playwright-graphql/coverage-reporter', {
             graphqlFilePath: './gql/graphql.ts', 
             coverageFilePath: './coverage/gql-coverage.log',
             htmlFilePath: './coverage/gql-coverage.html',
