+++ src/index.ts
@@ -1,2 +1,2 @@
 export { getSdkRequester } from './requester';
-export { coverageLogger } from './coverage';
+export { coverageLogger } from './coverage-reporter/coverageLogger';
